Rip out the original MortarConstraint
This commit compiles but the old tests that use MortarConstraint derived
objects fail because the input parameters are different. I was thinking
about going and re-doing these old meshes to create the subdomain and
boundary ID objects that the new constraint needs, but this caused me to
realize that the old mortar system works in 3D. I feel that we can't rip
out an object that supports higher-D mortar than the new implementation, so
I'm abandoning this tear-out for now.