ad_JxW is bad in step 9 of our tutorial
We're clearly not capturing the dependence of JxW on the displacements
correctly. This is really bizarre though because we test this in moose
test. I'm not sure what's different here. This seems to me to be more
motivation to template the FE object in libMesh. Code duplication
just seems very prone to error.