Add ping-ponging tests
This batch of tests illustrates the susceptiblity of pure
node-face mechanical contact algorithms to ping-ponging. In this
case the ping-ponging comes from oscillation of a slave node
between two different master faces. A hybrid scheme in which
LM residuals are computed at slave nodes while primal residuals
are computed on mortar integrals does not display oscillations.
This does not necessarily mean that the hybrid scheme is immune
to oscillations, but it at least performs better on this test
case.

Refs #10950