mesh: initialize right-elem properties correctly on boundaries
Although right-elem's are null on the boundary (they don't exist), we
still want to have some right-elem properties available for direct use
in the face info object to facilitate ghost elements for fv bcs.  So
along boundaries, set the right element to have the same volume.  The
right's centroid is reflected from the left cell centroid through the
face centroid.