Add MOOSE-generated mesh case for P2P1
This shows zeroth order convergence for the LM variable when the
slave side is coarser. John does not see this with his libMesh code. He
sees what you should see (e.g. O(h) to O(h^1.5)). So there is definitely
a bug in the mortar code. These will serve as good test cases for
ascertaining when the bug is fixed