Remove fi->isBoundary check in NSFVKernel
It appears that if a zero gradient condition is enforced when computing
cell gradients in MooseVariableFV that explicitly leaving off any
gradient contribution for the pressure in the RC interpolation is not
necessary. It is not immediately obvious to me why that works, but it's
definitely good from a generalization stand-point.