Verify INSFV in 1D
Observe second order convergence in the velocity but zeroth order
convergence in the pressure. Inspecting the pressure solution, we see
the checkerboard effect, which is undoubtedly to blame for the zeroth
order pressure convergence. Just a reminder that we see the checkerboard
because only grad_p matters in the solution and the grad_p
discretization does not have any dependence on the "current" cell, only
the left and right cells