Add test for 1D RZ RC with all required INSFV kernels
Demonstrates second order convergence in both variables. This test
includes an additional reaction kernel for the pressure. I believe that
this object is required in order for me to be able to use a Dirichlet
condition for the pressure, else the system would be singular. This may
not be true for an RC interpolation scheme though...