Demonstrate pure advective flow outlet condition
This shows second order convergence. Note that it is my conclusion that
it is impossible to enforce any non-dirichlet BC other than a zero
gradient BC and demonstrate second order convergence of the finite
volume method, at least in its current implementation. This is because
if we don't have a Dirichlet BC, then we assume a zero gradient in
`MooseVariableDataFV` in order to compute ghost values. Moreover if
there *is* a gradient and we're using something like an advection BC,
then we'll only see first order convergence because the solution used
for advection at the boundary is simply that of the cell center (again
we see that zero gradient assumption in action).