Verify 2D RC with channel-flow like zero-gradient bcs
I ended up having to add an FVDiffusion kernel to the mass equation in
order to observe optimal convergence. I don't have an answer for why
this is, but I *think* this is still a valid verification. I know that
I've verified the `NSFVKernel` in other MMS tests and if I change the
mms solution to have a non-zero gradient at a zero-gradient boundary,
then I no longer see optimal convergence. So it seems like everything is
verified...?

Ok, strike the above. I was able to get second order convergence for all
variables without using an FVDiffusion kernel in the mass equation.
The difference? Making sure that I was setting dirichlet conditions
on the top and bottom boundary for the *y* velocity, not the *x* velocity.
In a real problem, we're definitely setting the normal velocity to
be zero at the walls. We may or may not set the tangential velocity
to zero at the walls. I think that there is a fundamental problem with
the math if trying to set the normal gradient of the normal velocity *and*
the normal gradient of the pressure to zero at a wall boundary. I
haven't been able to work out why on paper, but that's my intuition.