Get correct mix of zero gradient/dirichlet bcs for 2d average
With the correct mix of zero gradient/dirichlet bcs, e.g. dirichlet
bc for normal velocity at the walls and zero gradient for the
tangential velocity at the walls (instead of the reverse which I
had formerly), there is no longer a need for an `FVDiffusion` kernel
in the mass equation in order to make the equation set nonsingular.
I believe you get a singularity if you try to set both the normal gradient
of the normal velocity and the normal gradient of the pressure to
zero on the same boundary. I am quite confident of this now actually.