Remove silly if statement in RC interpolation
Removing it doesn't affect the convergence of the interpolation scheme
and eye tests on the physics tests don't show a noticeable difference.
Morever, this if statement introduces discontinuities in the function
which are bad for a variety of reasons. A tangible one is it destroys
the ability to run PetscJacobianTesters