Create MooseMesh::determineUseDistributedMesh
This method is executed upon MooseMesh construction and whenever a user
calls MooseMesh::setParallelType. The addition of this method and its
execution in the MooseMesh constructor instead of in
MooseMesh::buildMeshBaseObject was necessary because we now add
geometric relationship managers before the mesh base object is created
when we're not using mesh generators. MooseApp checks MooseMesh for
whether the mesh is distributed when geometric relationship managers are
added, we we have to determine that information before those geometric
rms are added