Add API for creating specific libMesh mesh type
Right now BISON does a lot of manual creation of `ReplicatedMesh` in
their `MeshGenerators`. The problem with this is that it makes it
impossible for us to properly signal not to delete remote elements or to
attach appropriate ghosting functors if our simulation objects have
indicated that we need them. We should have an easy API for
`MeshGenerators` to construct specific types of libMesh meshes and would
then encourage generator developers to go through the `MooseMesh`
object, which will allow us to do our remote element deletion signalling
and attach appropriate relationship managers/ghosting functors. Of
course remote element management is not relevant for `ReplicatedMeshes`
but we want to anticipate possible `DistributedMesh` mesh generation in
the future.

Refs #14329