Demonstrate step change test with all terms in Euler momentum equation
The key here was the knowledge that for subsonic flows you can only
prescribe two conditions at the inlet and that you need to prescribe one
at the outlet. So now we prescribe pressure at the outlet and it works
like a beautiful charm. Note that the test added here uses the
"interpolate all components of the advection term the same way"
implementation. And this implementation seems to only work with 'average'
and does not work with upwind. You can also see the effects of 'average'
that we have seen before. E.g. an advected quantity will have almost the
same value across two elements. So this makes me think that we really
should never treat all components of an advection term the same way with
respect to interpolation. I think we probably should always be
averaging the advector (I believe including the advector in the upwind scheme is unstable
as illustrated by the fact that the test from this commit doesn't run to
completion because of convergence issues related to oscillations near
the velocity front. Just thinking about it intuitivity I feel like a
clang-bar effect in time could happen where we switch the winding
direction back and forth). And then we can either upwind or average the
advected quantity based on the user choice (e.g. what is currently done
in `FVMatAdvection`)