Implement canonical 1D/2D transient porosity step change tests
These tests average the advector and either upwind or average the
advected based on user specification. Two different advectors are used:
1. mass flux which will always be continuous
2. real velocity which will be continuous for compressible fluids
The advector chosen should always be continuous because attempting to
average discontinuous quantities leads to spurious oscillations. For
these tests if you look at the transient evolution you will see that the
more quantities are upwinded (e.g. if you use real velocity as the
advector instead of the mass flux), then the less you see the phenomena
where quantities are nearly continuous across two elements and then have
large jumps in between the two element segments. However, even when
using the real velocity as the advector, you still see some level of the
two-element behavior, presumably because the advector is averaged. As
commented in a previous commit a naive scheme of upwinding even the
advector is unstable, so averaging the advector and upwinding the
advected is the best we can do at least when considering just the naive
interpolation schemes of average and upwind. Probably with HLLC we would
do much better