Add a whole bunch of first-order BCs
These are first order because they the simply use the adjoining cell
centroid value for boundary information, e.g. we are using piecewise
constant data approximation as we interpolate from the cell center to
the face, e.g. we are not attempting to use any gradient information to
do a higher-order reconstruction of a boundary face value. This first
order approximation is especially fine because we are only using first
order information in our HLLC implementation