Some CNSFVHLLC refactoring
This work was started with the idea of creating special kernels and bcs
for discretizing sources within a HLLC setup, hence the creation of
these `Base` classes. Ultimately I'm unsure of whether the source
discretization based on the idea of characteristics (e.g. a
characteristic-based discretization of the source will often lead in
practice to upwinding the source) will go anywhere but this refactor may
be a good thing in its own right for other possible future development