Add regular grid examples showing change in mass with HLLC
The incoming mass flux is 1.28969 so that is what I would expect to see
for the value of `rho_u` in the domain. This is what I see when sources
are not present. However, when the source is present, I get a value of
approximately .91. This delta decreases as I refine the mesh