Add step change in porosity test with correct result
For sufficiently low Mach numbers a fluid is essentially incompressible
so when moving between different porosities (or flow areas analogously)
at low Mach number we should expect to see the *velocity* change and
*not* the density. This is what I see now that I'm correctly
incorporating the p * grad_eps term. This test is based on a test
provided me by Paolo. The results from PNSFV qualitatively match well
his results with Relap-5 in that:

- Big changes are in the particle velocity field - more or less
  inversely proportional to the change in porosity
- Very small changes in density/pressure

Paolo's relap-5 results show some definite trends for pressure/density
when the y-axis is really compressed. I don't currently see this with
the PNSFV implementation but I think I'm satisfied with PNSFV results
because they show that on the scale of the velocity change, the density
does not change (relatively speaking)