Add natural convection test using compressible NS
This relies on central differencing of the advection terms. For some
reason I cannot get this to work with the `PCNSFVLaxFriedrichs` objects.
My hypothesis is that it's because of way too much dissipation for these
rather low Mach number flows. This is something Josh and I have talked
about.

This commit includes tests for Rayleigh numbers of:

- 1.7e4
- 1.4e5