Add limiting to PCNSFVLaxFriedrichs
van-leer doesn't appear to work well when executed with a `Steady`
executioner. Depending on the number of elements, you can get singular
values (weird huh?). However, it converges well after adding dot
computation to `PorousPrimitiveVarMaterial` and running with
`Transient`. I should qualify runs well: it only appears to converge for
my heated-channel case if I run with `two_term_boundary_expansion =
false`. With this one-term-expansion van-leer solution the density,
superficial velocity, and temperature profiles look good. However, there
are artifacts at the boundaries for the pressure solution. ~My next
attempt will be to see about interpolating different quantities to the
faces. The variable set I would try is: rho, sup_rho_u, and T because
those are the fields that they interpolate in the Greenshields paper~

Scratch that for now... because there are no artifacts in the upwind or
central-difference limited schemes I'm inclined to think there are no
issues due to the Lax-Friedrichs kernels nor with the porous primitive
material (e.g. no issue with the choice of variable set). I'm more
inclined to think the issue is with the van-leer limiting itself,
particulary in the limiting of the pressure solution at the boundaries.
So next thing I'm going to try probably is printing out the beta values
for the mass/pressure equation and see whether anything looks suspicious

Yet another note: toggling off the pressure term I don't understand in
the energy equation in the Lax-Friedrichs objects doesn't change the
non-monotone behavior in the pressure solution. (It moves the whole solution
vertically but that is it)

Update: 5/6/21

I tried adding dirichlet boundary conditions to improve the quality of
the computed cell center gradients adjacent to boundaries. With these
plux `two_term_boundary_expansion` I can get a monotone pressure
solution with small enough meshes using vanLeer limiting. However, I now see huge gradients in
the temperature solution at the right boundary