Banish manual manipulation of ADReals in KT sims
Anytime you see a `.value()` or a `.derivatives()` you should be
alarmed. It means that you are almost certainly going to mess up the
Jacobian. Note to future Alex: don't write or allow any code paths
present in your simulation to include manual manipulation of ADReals.
This was screwing up my CNSFV sims because I wasn't accounting for the
fact that things like dp_dv *also* depend on the nonlinear degrees of
freedom