Add continuously varying porosity test
This test stongly suggests that all my previous step porosity change
tests are wrong. The new test matches the old velocity profiles roughly
but the pressure and density profiles are totally different. The density
profiles for the new continuous test I definitely believe more
considering the MMS tests I have for the continuous case and because of
the relap-5 results Paolo shared with me. So I need to figure out whats
wrong with the porosity jump object