Add "strong" way to enforce boundary conditions
This isn't actually strong enforcement as in strongly setting a degree
of freedom to a given value. Instead it is directly setting fluxes at
boundaries based only off of information provided by the user, e.g. it
uses as little interior information as possible. This is different from
the characterisic-based BCs that use a lot of information from the
interior to form the boundary flux (with the idea that you are
propagating waves both into and out of the domain). However, for physics
that we care about we want our boundary conditions honored so we are
going to discard a characterisic-based implementation that is not
ensured to give us the boundary fluxes we want