Use ExplicitSSPRungeKutta in CNSFV explicit solves
We weren't using the tvd-rk2 time integrator correctly (we weren't
setting `implicit = false` for non-time kernels) nor is it clear to me
that the tvd-rk2 time integrator is actually implemented correctly. You
can see https://github.com/idaholab/moose/pull/6605 for a three-point
plot of error convergence with slope of 2.65.